api/requirements.txt
fastapi==0.87.0
uvicorn==0.20.0
kafka-python==2.0.2
clickhouse-connect==0.4.5 
